Example prompt that shows the current memory usage of the R process

Example prompt that shows the current memory usage of the R process

prompt_mem(...)

memory_usage()

Arguments

...

Ignored.

Value

prompt_mem() returns the formatted prompt in a string.

memory_usage() returns a list with entries:

bytes: the number of bytes of memory the current process uses. This is the 'Resident Set Size', see ps::ps_memory_info().

formatted: string that formats bytes nicely, with the appropriate unit.

total: Total physical memory. See ps::ps_system_memory().

avail: the memory that can be given instantly to processes without the system going into swap. See ps::ps_system_memory().

percent: Percentage of memory that is taken. See ps::ps_system_memory().

Details

prompt_mem() is a simple example prompt that shows the physical memory allocated by the current process.

memory_usage() is a utility function that shows memory information about the current R process and the system. You can use it to create a custom prompt.

See also

Examples

cat(prompt_mem())
#> 142.41 MB ❯ 
memory_usage()
#> $bytes
#> [1] 142675968
#> 
#> $formatted
#> [1] "142.68 MB"
#> 
#> $total
#> [1] 7268679680
#> 
#> $avail
#> [1] 6218092544
#> 
#> $percent
#> [1] 14.45362
#>