Set and control the prompt

Set and control the prompt

set_prompt(value)

suspend()

restore()

toggle()

Arguments

value

A character string for a static prompt, or a function that is called after the evaluation every expression typed at the R prompt. The function should always return a character scalar.

Value

No return value, called for side effects.

Details

Function update_prompt() is used to replace the default R prompt with a custom prompt. A custom prompt can be disabled with suspend() and then re-enable with restore(). Function toggle() toggles between the two.